admin/be/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# ML/NLP (for admin features)
openai==1.3.5
transformers==4.35.2
torch==2.8.0
pandas==2.1.3
numpy==1.26.2
accelerate==1.11.0
underthesea==8.3.0

# Utilities
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

pydantic[email]==2.5.2
